+++ src/components/Table/TdCheckbox.tsx
@@ -1,12 +1,11 @@
-import React, { ReactNode, forwardRef } from 'react'
-import styled, { css } from 'styled-components'
+import React, { ReactNode, forwardRef, useMemo } from 'react'
+import { tv } from 'tailwind-variants'
 
-import { Theme, useTheme } from '../../hooks/useTheme'
 import { CheckBox, Props as CheckBoxProps } from '../CheckBox'
 import { Center } from '../Layout'
 import { VisuallyHiddenText } from '../VisuallyHiddenText'
 
-import { Td as shrTd } from './Td'
+import { Td } from './Td'
 
 type Props = {
   /** 値を特定するための行 id */
@@ -15,35 +14,33 @@ type Props = {
   children?: ReactNode
 }
 
+const tdCheckbox = tv({
+  slots: {
+    inner: 'shr-absolute shr-inset-0 [&:not(:has([disabled]))]:shr-cursor-pointer',
+    wrapper: 'shr-relative shr-w-[theme(fontSize.base)] [&]:shr-p-0.75',
+  },
+})
+
 export const TdCheckbox = forwardRef<HTMLInputElement, Omit<CheckBoxProps, keyof Props> & Props>(
   ({ 'aria-labelledby': ariaLabelledby, children, className, ...others }, ref) => {
-    const theme = useTheme()
+    const { wrapperStyle, innerStyle } = useMemo(() => {
+      const { wrapper, inner } = tdCheckbox()
+      return {
+        wrapperStyle: wrapper({ className }),
+        innerStyle: inner(),
+      }
+    }, [className])
+
     return (
       // Td に必要な属性やイベントは不要
-      <Td className={className} $themes={theme}>
-        <LabelCenter>
+      <Td className={wrapperStyle}>
+        <Center as="label" verticalCentering className={innerStyle}>
           {/* 使う側で lint をかけるため無効化 */}
           {/* eslint-disable-next-line smarthr/a11y-input-has-name-attribute  */}
           <CheckBox {...others} ref={ref} aria-labelledby={ariaLabelledby} />
           {children && <VisuallyHiddenText>{children}</VisuallyHiddenText>}
-        </LabelCenter>
+        </Center>
       </Td>
     )
   },
 )
-
-const Td = styled(shrTd)<{ $themes: Theme }>`
-  ${({ $themes: { fontSize, space } }) => css`
-    position: relative;
-    padding: ${space(0.75)};
-    width: ${fontSize.M};
-  `}
-`
-const LabelCenter = styled(Center).attrs({ forwardedAs: 'label', verticalCentering: true })`
-  position: absolute;
-  inset: 0;
-
-  &:not(:has([disabled])) {
-    cursor: pointer;
-  }
-`